refactor(xtask): 重构架构相关操作
将操作架构相关的 linux rootfs 的操作与架构定义分离，放到独立的模块中
