+++ xtask/CHANGLOG.md
@@ -2,26 +2,30 @@
 
 最新的更新将出现在最上方。
 
-## 20220604(YdrMaster)
+## 20220609 (YdrMaster)
+
+- 重构架构相关操作，将操作架构相关的 linux rootfs 的操作与架构定义分离，放到独立的模块中。
+
+## 20220604 (YdrMaster)
 
 - Thanks to 国防科大的工作，现在 zCore 支持 aarch64，xtask 各个架构相关指令现在也都支持 `--arch=aarch64`；
 
-## 20220521(YdrMaster)
+## 20220521 (YdrMaster)
 
 - 适用于 libos 的 musl libc so 文件不再通过 git lfs 获取；
 
-## 20220513(YdrMaster)
+## 20220513 (YdrMaster)
 
 - 选择架构现在是一个参数而不是子命令，例如 `cargo rootfs --arch x86_64`；
 - 增加 `asm` 指令将指定参数编译的内核反汇编到文件；
 - 增加 `qemu` 指令在 QEMU 中启动 zCore（目前仅支持 RiscV64+Linux），可配置 SMP，可配置连接 gdb；
 - 增加 `gdb` 指令启动 gdb 并连接指定端口（目前仅支持 RiscV64）。
 
-## 20220512(YdrMaster)
+## 20220512 (YdrMaster)
 
 `cargo check-style` 现在会依 CI/build 的方式工作。
 
-## 20220511(YdrMaster)
+## 20220511 (YdrMaster)
 
 ### 目录结构定义
 
@@ -43,13 +47,13 @@
 
 - 使用 `std::os::unix::fs::symlink` 建立符号链接，不再依赖 `ln` 应用程序；
 
-## 20220506(YdrMaster)
+## 20220506 (YdrMaster)
 
 顶层的 Makefile 已经尽量迁移到 rust，并在子项目 README.md 中更新了子命令说明。
 
 计划提起一次 PR。
 
-## 20220504(YdrMaster)
+## 20220504 (YdrMaster)
 
 初步的计划是先尽量将 Makefile 转化为类型安全且更有可能工程化结构化的 Rust xtask。
 尤其是要将 zCore 目录内外的两个 Makefile 合并。
